Add more flexible solution checking logic
